Create useAuthHandlers and useChatRoomHandlers and fix User typing
